Uploaded "Guess The Number" Python Application
This application has the user guess a number between 1 to 1000 and counts the amount of guesses of the user.
It will tell the user if the value guessed is too low or too high and ask to try again. If the user guesses the number
correct, it will ask the user if they want to play again.